+++ runtime/vm/compiler/runtime_offsets_extracted.h
@@ -195,9 +195,9 @@ static constexpr dart::compiler::target::word SubtypeTestCache_cache_offset = 4;
 static constexpr dart::compiler::target::word
     Thread_AllocateArray_entry_point_offset = 292;
 static constexpr dart::compiler::target::word Thread_active_exception_offset =
-    624;
-static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
     628;
+static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
+    632;
 static constexpr dart::compiler::target::word
     Thread_array_write_barrier_code_offset = 120;
 static constexpr dart::compiler::target::word
@@ -212,7 +212,7 @@ static constexpr dart::compiler::target::word
     Thread_call_to_runtime_entry_point_offset = 208;
 static constexpr dart::compiler::target::word
     Thread_call_to_runtime_stub_offset = 140;
-static constexpr dart::compiler::target::word Thread_dart_stream_offset = 660;
+static constexpr dart::compiler::target::word Thread_dart_stream_offset = 664;
 static constexpr dart::compiler::target::word Thread_optimize_entry_offset =
     236;
 static constexpr dart::compiler::target::word Thread_optimize_stub_offset = 164;
@@ -228,7 +228,7 @@ static constexpr dart::compiler::target::word Thread_end_offset = 64;
 static constexpr dart::compiler::target::word
     Thread_enter_safepoint_stub_offset = 188;
 static constexpr dart::compiler::target::word Thread_execution_state_offset =
-    644;
+    648;
 static constexpr dart::compiler::target::word
     Thread_exit_safepoint_stub_offset = 192;
 static constexpr dart::compiler::target::word
@@ -248,7 +248,7 @@ static constexpr dart::compiler::target::word Thread_float_not_address_offset =
 static constexpr dart::compiler::target::word
     Thread_float_zerow_address_offset = 288;
 static constexpr dart::compiler::target::word Thread_global_object_pool_offset =
-    632;
+    636;
 static constexpr dart::compiler::target::word
     Thread_interpret_call_entry_point_offset = 256;
 static constexpr dart::compiler::target::word
@@ -283,11 +283,11 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_object_null_offset = 104;
 static constexpr dart::compiler::target::word
     Thread_predefined_symbols_address_offset = 260;
-static constexpr dart::compiler::target::word Thread_resume_pc_offset = 636;
+static constexpr dart::compiler::target::word Thread_resume_pc_offset = 640;
 static constexpr dart::compiler::target::word
-    Thread_saved_shadow_call_stack_offset = 640;
+    Thread_saved_shadow_call_stack_offset = 644;
 static constexpr dart::compiler::target::word Thread_safepoint_state_offset =
-    648;
+    652;
 static constexpr dart::compiler::target::word
     Thread_slow_type_test_stub_offset = 180;
 static constexpr dart::compiler::target::word Thread_stack_limit_offset = 36;
@@ -318,7 +318,7 @@ static constexpr dart::compiler::target::word
     Thread_write_barrier_entry_point_offset = 200;
 static constexpr dart::compiler::target::word Thread_write_barrier_mask_offset =
     48;
-static constexpr dart::compiler::target::word Thread_callback_code_offset = 652;
+static constexpr dart::compiler::target::word Thread_callback_code_offset = 656;
 static constexpr dart::compiler::target::word TimelineStream_enabled_offset = 8;
 static constexpr dart::compiler::target::word TwoByteString_data_offset = 12;
 static constexpr dart::compiler::target::word Type_arguments_offset = 16;
@@ -351,8 +351,8 @@ static constexpr dart::compiler::target::word
     Code_function_entry_point_offset[] = {4, 8};
 static constexpr dart::compiler::target::word
     Thread_write_barrier_wrappers_thread_offset[] = {
-        588, 592, 596, 600, 604, -1, 608, 612,
-        616, 620, -1,  -1,  -1,  -1, -1,  -1};
+        592, 596, 600, 604, 608, -1, 612, 616,
+        620, 624, -1,  -1,  -1,  -1, -1,  -1};
 static constexpr dart::compiler::target::word Array_header_size = 12;
 static constexpr dart::compiler::target::word Context_header_size = 12;
 static constexpr dart::compiler::target::word Double_InstanceSize = 16;
@@ -556,9 +556,9 @@ static constexpr dart::compiler::target::word SubtypeTestCache_cache_offset = 8;
 static constexpr dart::compiler::target::word
     Thread_AllocateArray_entry_point_offset = 568;
 static constexpr dart::compiler::target::word Thread_active_exception_offset =
-    1248;
-static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
     1256;
+static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
+    1264;
 static constexpr dart::compiler::target::word
     Thread_array_write_barrier_code_offset = 224;
 static constexpr dart::compiler::target::word
@@ -573,7 +573,7 @@ static constexpr dart::compiler::target::word
     Thread_call_to_runtime_entry_point_offset = 400;
 static constexpr dart::compiler::target::word
     Thread_call_to_runtime_stub_offset = 264;
-static constexpr dart::compiler::target::word Thread_dart_stream_offset = 1320;
+static constexpr dart::compiler::target::word Thread_dart_stream_offset = 1328;
 static constexpr dart::compiler::target::word Thread_optimize_entry_offset =
     456;
 static constexpr dart::compiler::target::word Thread_optimize_stub_offset = 312;
@@ -589,7 +589,7 @@ static constexpr dart::compiler::target::word Thread_end_offset = 128;
 static constexpr dart::compiler::target::word
     Thread_enter_safepoint_stub_offset = 360;
 static constexpr dart::compiler::target::word Thread_execution_state_offset =
-    1288;
+    1296;
 static constexpr dart::compiler::target::word
     Thread_exit_safepoint_stub_offset = 368;
 static constexpr dart::compiler::target::word
@@ -609,7 +609,7 @@ static constexpr dart::compiler::target::word Thread_float_not_address_offset =
 static constexpr dart::compiler::target::word
     Thread_float_zerow_address_offset = 560;
 static constexpr dart::compiler::target::word Thread_global_object_pool_offset =
-    1264;
+    1272;
 static constexpr dart::compiler::target::word
     Thread_interpret_call_entry_point_offset = 496;
 static constexpr dart::compiler::target::word
@@ -644,11 +644,11 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_object_null_offset = 192;
 static constexpr dart::compiler::target::word
     Thread_predefined_symbols_address_offset = 504;
-static constexpr dart::compiler::target::word Thread_resume_pc_offset = 1272;
+static constexpr dart::compiler::target::word Thread_resume_pc_offset = 1280;
 static constexpr dart::compiler::target::word
-    Thread_saved_shadow_call_stack_offset = 1280;
+    Thread_saved_shadow_call_stack_offset = 1288;
 static constexpr dart::compiler::target::word Thread_safepoint_state_offset =
-    1296;
+    1304;
 static constexpr dart::compiler::target::word
     Thread_slow_type_test_stub_offset = 344;
 static constexpr dart::compiler::target::word Thread_stack_limit_offset = 72;
@@ -680,7 +680,7 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_write_barrier_mask_offset =
     96;
 static constexpr dart::compiler::target::word Thread_callback_code_offset =
-    1304;
+    1312;
 static constexpr dart::compiler::target::word TimelineStream_enabled_offset =
     16;
 static constexpr dart::compiler::target::word TwoByteString_data_offset = 16;
@@ -714,8 +714,8 @@ static constexpr dart::compiler::target::word
     Code_function_entry_point_offset[] = {8, 16};
 static constexpr dart::compiler::target::word
     Thread_write_barrier_wrappers_thread_offset[] = {
-        1160, 1168, 1176, 1184, -1,   -1,   1192, 1200,
-        1208, 1216, 1224, -1,   1232, 1240, -1,   -1};
+        1168, 1176, 1184, 1192, -1,   -1,   1200, 1208,
+        1216, 1224, 1232, -1,   1240, 1248, -1,   -1};
 static constexpr dart::compiler::target::word Array_header_size = 24;
 static constexpr dart::compiler::target::word Context_header_size = 24;
 static constexpr dart::compiler::target::word Double_InstanceSize = 16;
@@ -917,9 +917,9 @@ static constexpr dart::compiler::target::word SubtypeTestCache_cache_offset = 4;
 static constexpr dart::compiler::target::word
     Thread_AllocateArray_entry_point_offset = 292;
 static constexpr dart::compiler::target::word Thread_active_exception_offset =
-    588;
-static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
     592;
+static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
+    596;
 static constexpr dart::compiler::target::word
     Thread_array_write_barrier_code_offset = 120;
 static constexpr dart::compiler::target::word
@@ -934,7 +934,7 @@ static constexpr dart::compiler::target::word
     Thread_call_to_runtime_entry_point_offset = 208;
 static constexpr dart::compiler::target::word
     Thread_call_to_runtime_stub_offset = 140;
-static constexpr dart::compiler::target::word Thread_dart_stream_offset = 624;
+static constexpr dart::compiler::target::word Thread_dart_stream_offset = 628;
 static constexpr dart::compiler::target::word Thread_optimize_entry_offset =
     236;
 static constexpr dart::compiler::target::word Thread_optimize_stub_offset = 164;
@@ -950,7 +950,7 @@ static constexpr dart::compiler::target::word Thread_end_offset = 64;
 static constexpr dart::compiler::target::word
     Thread_enter_safepoint_stub_offset = 188;
 static constexpr dart::compiler::target::word Thread_execution_state_offset =
-    608;
+    612;
 static constexpr dart::compiler::target::word
     Thread_exit_safepoint_stub_offset = 192;
 static constexpr dart::compiler::target::word
@@ -970,7 +970,7 @@ static constexpr dart::compiler::target::word Thread_float_not_address_offset =
 static constexpr dart::compiler::target::word
     Thread_float_zerow_address_offset = 288;
 static constexpr dart::compiler::target::word Thread_global_object_pool_offset =
-    596;
+    600;
 static constexpr dart::compiler::target::word
     Thread_interpret_call_entry_point_offset = 256;
 static constexpr dart::compiler::target::word
@@ -1005,11 +1005,11 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_object_null_offset = 104;
 static constexpr dart::compiler::target::word
     Thread_predefined_symbols_address_offset = 260;
-static constexpr dart::compiler::target::word Thread_resume_pc_offset = 600;
+static constexpr dart::compiler::target::word Thread_resume_pc_offset = 604;
 static constexpr dart::compiler::target::word
-    Thread_saved_shadow_call_stack_offset = 604;
+    Thread_saved_shadow_call_stack_offset = 608;
 static constexpr dart::compiler::target::word Thread_safepoint_state_offset =
-    612;
+    616;
 static constexpr dart::compiler::target::word
     Thread_slow_type_test_stub_offset = 180;
 static constexpr dart::compiler::target::word Thread_stack_limit_offset = 36;
@@ -1040,7 +1040,7 @@ static constexpr dart::compiler::target::word
     Thread_write_barrier_entry_point_offset = 200;
 static constexpr dart::compiler::target::word Thread_write_barrier_mask_offset =
     48;
-static constexpr dart::compiler::target::word Thread_callback_code_offset = 616;
+static constexpr dart::compiler::target::word Thread_callback_code_offset = 620;
 static constexpr dart::compiler::target::word TimelineStream_enabled_offset = 8;
 static constexpr dart::compiler::target::word TwoByteString_data_offset = 12;
 static constexpr dart::compiler::target::word Type_arguments_offset = 16;
@@ -1274,9 +1274,9 @@ static constexpr dart::compiler::target::word SubtypeTestCache_cache_offset = 8;
 static constexpr dart::compiler::target::word
     Thread_AllocateArray_entry_point_offset = 568;
 static constexpr dart::compiler::target::word Thread_active_exception_offset =
-    1336;
-static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
     1344;
+static constexpr dart::compiler::target::word Thread_active_stacktrace_offset =
+    1352;
 static constexpr dart::compiler::target::word
     Thread_array_write_barrier_code_offset = 224;
 static constexpr dart::compiler::target::word
@@ -1291,7 +1291,7 @@ static constexpr dart::compiler::target::word
     Thread_call_to_runtime_entry_point_offset = 400;
 static constexpr dart::compiler::target::word
     Thread_call_to_runtime_stub_offset = 264;
-static constexpr dart::compiler::target::word Thread_dart_stream_offset = 1408;
+static constexpr dart::compiler::target::word Thread_dart_stream_offset = 1416;
 static constexpr dart::compiler::target::word Thread_optimize_entry_offset =
     456;
 static constexpr dart::compiler::target::word Thread_optimize_stub_offset = 312;
@@ -1307,7 +1307,7 @@ static constexpr dart::compiler::target::word Thread_end_offset = 128;
 static constexpr dart::compiler::target::word
     Thread_enter_safepoint_stub_offset = 360;
 static constexpr dart::compiler::target::word Thread_execution_state_offset =
-    1376;
+    1384;
 static constexpr dart::compiler::target::word
     Thread_exit_safepoint_stub_offset = 368;
 static constexpr dart::compiler::target::word
@@ -1327,7 +1327,7 @@ static constexpr dart::compiler::target::word Thread_float_not_address_offset =
 static constexpr dart::compiler::target::word
     Thread_float_zerow_address_offset = 560;
 static constexpr dart::compiler::target::word Thread_global_object_pool_offset =
-    1352;
+    1360;
 static constexpr dart::compiler::target::word
     Thread_interpret_call_entry_point_offset = 496;
 static constexpr dart::compiler::target::word
@@ -1362,11 +1362,11 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_object_null_offset = 192;
 static constexpr dart::compiler::target::word
     Thread_predefined_symbols_address_offset = 504;
-static constexpr dart::compiler::target::word Thread_resume_pc_offset = 1360;
+static constexpr dart::compiler::target::word Thread_resume_pc_offset = 1368;
 static constexpr dart::compiler::target::word
-    Thread_saved_shadow_call_stack_offset = 1368;
+    Thread_saved_shadow_call_stack_offset = 1376;
 static constexpr dart::compiler::target::word Thread_safepoint_state_offset =
-    1384;
+    1392;
 static constexpr dart::compiler::target::word
     Thread_slow_type_test_stub_offset = 344;
 static constexpr dart::compiler::target::word Thread_stack_limit_offset = 72;
@@ -1398,7 +1398,7 @@ static constexpr dart::compiler::target::word
 static constexpr dart::compiler::target::word Thread_write_barrier_mask_offset =
     96;
 static constexpr dart::compiler::target::word Thread_callback_code_offset =
-    1392;
+    1400;
 static constexpr dart::compiler::target::word TimelineStream_enabled_offset =
     16;
 static constexpr dart::compiler::target::word TwoByteString_data_offset = 16;
@@ -1432,9 +1432,9 @@ static constexpr dart::compiler::target::word
     Code_function_entry_point_offset[] = {8, 16};
 static constexpr dart::compiler::target::word
     Thread_write_barrier_wrappers_thread_offset[] = {
-        1160, 1168, 1176, 1184, 1192, 1200, 1208, 1216, 1224, 1232, 1240,
-        1248, 1256, 1264, 1272, -1,   -1,   -1,   -1,   1280, 1288, 1296,
-        1304, 1312, 1320, 1328, -1,   -1,   -1,   -1,   -1,   -1};
+        1168, 1176, 1184, 1192, 1200, 1208, 1216, 1224, 1232, 1240, 1248,
+        1256, 1264, 1272, 1280, -1,   -1,   -1,   -1,   1288, 1296, 1304,
+        1312, 1320, 1328, 1336, -1,   -1,   -1,   -1,   -1,   -1};
 static constexpr dart::compiler::target::word Array_header_size = 24;
 static constexpr dart::compiler::target::word Context_header_size = 24;
 static constexpr dart::compiler::target::word Double_InstanceSize = 16;