+++ runtime/vm/runtime_entry.cc
@@ -3025,9 +3025,16 @@ DEFINE_RUNTIME_ENTRY(UpdateFieldCid, 2) {
 #endif
 }
 
+DEFINE_RUNTIME_ENTRY(InitInstanceField, 2) {
+  const Instance& instance = Instance::CheckedHandle(zone, arguments.ArgAt(0));
+  const Field& field = Field::CheckedHandle(zone, arguments.ArgAt(1));
+  const Error& result = Error::Handle(zone, field.InitializeInstance(instance));
+  ThrowIfError(result);
+}
+
 DEFINE_RUNTIME_ENTRY(InitStaticField, 1) {
   const Field& field = Field::CheckedHandle(zone, arguments.ArgAt(0));
-  const Error& result = Error::Handle(zone, field.Initialize());
+  const Error& result = Error::Handle(zone, field.InitializeStatic());
   ThrowIfError(result);
 }