+++ runtime/vm/runtime_entry_list.h
@@ -52,6 +52,7 @@ namespace dart {
   V(DeoptimizeMaterialize)                                                     \
   V(RewindPostDeopt)                                                           \
   V(UpdateFieldCid)                                                            \
+  V(InitInstanceField)                                                         \
   V(InitStaticField)                                                           \
   V(CompileFunction)                                                           \
   V(CompileInterpretedFunction)                                                \